refactor(stats): split CommunityStats into smaller components and extract data to constants
